# ai-dev-input/02_api_contracts/engines.openapi.yaml
openapi: 3.0.3
info:
  title: Specialty Care Pathway Engines API
  version: 1.0.0
  description: 专科诊疗路径平台三大引擎核心接口契约
servers:
  - url: /api
...
paths:
  /pathways:
    get:
      tags: [PathwayEngine]
      summary: 查询路径定义列表
      operationId: listPathways
      responses:
        '200':
          description: 路径定义列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
    post:
      tags: [PathwayEngine]
      summary: 创建路径定义草稿
      operationId: createPathway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathwayConfig'
      responses:
        '200':
          description: 创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
  /pathways/{pathwayCode}:
    get:
      tags: [PathwayEngine]
      summary: 查询路径定义和配置原文
      operationId: getPathway
      parameters:
        - $ref: '#/components/parameters/PathwayCode'
        - name: versionNo
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 路径定义详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
  /pathways/{pathwayCode}/publish:
    post:
      tags: [PathwayEngine]
      summary: 发布路径版本
      operationId: publishPathway
      parameters:
        - $ref: '#/components/parameters/PathwayCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_no, approved_by]
              properties:
                version_no:
                  type: string
                approved_by:
                  type: string
                comment:
                  type: string
      responses:
        '200':
          description: 发布结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
  /patient-pathways/candidates:
    post:
      tags: [PathwayEngine]
      summary: 候选路径识别
      operationId: findCandidatePathways
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientContext'
      responses:
        '200':
          description: 候选路径列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RecommendationCard'
  /patient-pathways/admit:
    post:
      tags: [PathwayEngine]
      summary: 医生确认患者入径
      description: |
        组织上下文可通过 Header/Query/Body 传入；Body 中的 tenant_id/group_code/hospital_code/campus_code/site_code/department_code 优先覆盖 Header/Query。
        返回的患者路径实例会带组织字段，后续路径实例、变异、质控和审计查询可按组织维度过滤。
      operationId: admitPatientPathway
      parameters:
        - $ref: '#/components/parameters/HeaderTenantId'
        - $ref: '#/components/parameters/HeaderGroupCode'
        - $ref: '#/components/parameters/HeaderHospitalCode'
        - $ref: '#/components/parameters/HeaderCampusCode'
        - $ref: '#/components/parameters/HeaderSiteCode'
        - $ref: '#/components/parameters/HeaderDepartmentCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id, encounter_id, pathway_code, version_no, doctor_id]
              properties:
                patient_id:
                  type: string
                encounter_id:
                  type: string
                pathway_code:
                  type: string
                version_no:
                  type: string
                doctor_id:
                  type: string
                recommendation_id:
                  type: string
                tenant_id:
                  type: string
                group_code:
                  type: string
                hospital_code:
                  type: string
                campus_code:
                  type: string
                site_code:
                  type: string
                department_code:
                  type: string
      responses:
        '200':
          description: 患者路径实例
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientPathwayInstance'
  /patient-pathways/{instanceId}/nodes/{nodeCode}/complete:
    post:
      tags: [PathwayEngine]
      summary: 完成路径节点
      operationId: completePathwayNode
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
        - name: nodeCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operator_id:
                  type: string
                completion_reason:
                  type: string
                evidence:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 完成结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
...
components:
  parameters:
    PathwayCode:
      name: pathwayCode
      in: path
      required: true
      schema:
        type: string
    HeaderTenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: 多租户隔离标识；未提供时默认 default。
      schema: {type: string}
    HeaderGroupCode:
      name: X-Group-Code
      in: header
      required: false
      schema: {type: string}
    HeaderHospitalCode:
      name: X-Hospital-Code
      in: header
      required: false
      schema: {type: string}
    HeaderCampusCode:
      name: X-Campus-Code
      in: header
      required: false
      schema: {type: string}
    HeaderSiteCode:
      name: X-Site-Code
      in: header
      required: false
      schema: {type: string}
    HeaderDepartmentCode:
      name: X-Department-Code
      in: header
      required: false
      schema: {type: string}
  schemas:
    ApiResult:
      type: object
      required: [success, code, message, trace_id]
      properties:
        success:
          type: boolean
        code:
          type: string
          example: "0"
        message:
          type: string
          example: success
        trace_id:
          type: string
        data:
          nullable: true
...
    PatientContext:
      type: object
      required: [patient, encounter, facts]
      properties:
        patient:
          type: object
          additionalProperties: true
        encounter:
          type: object
          additionalProperties: true
        facts:
          type: object
          additionalProperties: true
    PathwayConfig:
      type: object
      required: [pathway_code, pathway_name, version, stages]
      properties:
        pathway_code:
          type: string
        pathway_name:
          type: string
        version:
          type: string
        stages:
          type: array
          items:
            type: object
            additionalProperties: true
...
    RecommendationCard:
      type: object
      properties:
        recommendation_id:
          type: string
        patient_id:
          type: string
        encounter_id:
          type: string
        target_code:
          type: string
        target_name:
          type: string
        score:
          type: number
        confidence:
          type: string
        supporting_facts:
          type: array
          items:
            type: object
        suggested_actions:
          type: array
          items:
            type: string
    PatientPathwayInstance:
      type: object
      properties:
        instance_id:
          type: string
        patient_id:
          type: string
        encounter_id:
          type: string
        pathway_code:
          type: string
        version_no:
          type: string
        status:
          type: string
        current_node_code:
          type: string
        tenant_id:
          type: string
        group_code:
          type: string
        hospital_code:
          type: string
        campus_code:
          type: string
        site_code:
          type: string
        department_code:
          type: string
        scope_level:
          type: string
        scope_code:
          type: string
        org_source:
          type: string
...